refactor(settings): extract breakpoints, category state, section header, and pickers
Pure refactoring; no visible behavior change.

- Add SettingsBreakpoints.masterDetail (600.0) as the single source for
  the compact/two-pane layout split.
- Add SettingsCategory enum + SettingsCategoryNotifier +
  settingsCategoryProvider (non-autoDispose, initial display).
  labelKey/icon are intentionally deferred to the l10n commit (P1-C4).
- Make _SectionHeader public as SettingsSectionHeader
  (toUpperCase removal and Semantics land with the l10n commit).
- Move all _show* dialog helpers into settings/pickers/ (11 files)
  with their original signatures, scoping the deprecated
  RadioListTile ignore to the 6 files that use it.
