+++ lib/screens/settings/widgets/settings_section_header.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+import '../../../theme/design_colors.dart';
+
+/// 設定画面のセクション（グループ）見出し。
+///
+/// P1-C4 で toUpperCase 廃止 + Semantics(header: true) が追加される（F-2）。
+class SettingsSectionHeader extends StatelessWidget {
+  final String title;
+
+  const SettingsSectionHeader({super.key, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(16, 16, 16, 8),
+      child: Text(
+        title.toUpperCase(),
+        style: GoogleFonts.spaceGrotesk(
+          fontSize: 11,
+          fontWeight: FontWeight.w700,
+          color: isDark ? DesignColors.textMuted : DesignColors.textMutedLight,
+          letterSpacing: 1.5,
+        ),
+      ),
+    );
+  }
+}
